+++ mdx-components.tsx
@@ -1,0 +1,9 @@
+import type { ComponentPropsWithoutRef } from 'react';
+
+type Components = {
+  [key: string]: React.ComponentType<ComponentPropsWithoutRef<any>>;
+};
+
+export function useMDXComponents(components: Components): Components {
+  return { ...components };
+}